Run tests with 32-bit php on travis.
The serializer contains 32-bit specific code and code dependent on
pointer lengths, and this should be tested.

`gcc -m32 ...` (ci/gcc-32.sh) builds 32-bit versions of binaries, including php.
(It will test the case where SIZEOF_ZEND_LONG == 4)
The 64-bit travis machines are capable of running those binaries.

This takes 1 minute to compile, so a travis caching step was added.
This has multiple layers of cache, trying to first specify the most
specific cache, then a more generic one for the whole project.
(The cache may have a different php installation, which is given a
different name so it can be detected and removed)

Based on https://github.com/runkit7/runkit7/blob/master/.travis.yml in
another project I work on.

Install CC before replacing it with a shell script for 32-bit
Add the flag --enable-json for php >= 7.0.10